requirements.txt
appnope==0.1.0
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.1.2
enum34==1.1.6
Flask==0.12.2
functools32==3.2.3.post2
gunicorn==19.7.1
idna==2.6
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.0
matplotlib==2.1.0
nbformat==4.4.0
numpy==1.13.3
oauthlib==2.0.6
pandas==0.21.0
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
plotly==2.2.3
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pymongo==3.6.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-igraph==0.7.1.post6
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
scandir==1.6
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
subprocess32==3.2.7
traitlets==4.3.2
twython==3.6.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
